# API-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebShop24",
    "description": "This is the best WebShop you can find.\n",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://petstore.swagger.io/v2"
    }
  ],
...
  "paths": {
...
    "/order": {
      "post": {
        "tags": [
          "order"
        ],
        "summary": "Create a new order",
        "operationId": "createOrder",
        "requestBody": {
          "$ref": "#/components/requestBodies/Order"
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "buyer": "cWGtlAnhcsVViidg",
                  "articles": [{
                    "productID": "cWGtlAnhcsVViidg",
                    "quantity": 2
                  }
                  ],
                  "totalAmount": 790.01,
                  "seller":"ofqRrrA7womznAeY",
                  "updatedAt":"2024-01-07T17:15:24.625Z",
                  "orderDate": "2023-12-01T14:30:00Z",
                  "status": "placed"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "order"
        ],
        "summary": "Get all orders",
        "operationId": "getOrders",
        "responses": {
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Orders returned",
            "content": {
              "application/json": {
                "example": [{
                  "_id": "cWGtlAnhcsVViidg",
                  "buyer": "cWGtlAnhcsVViidg",
                  "articles": [{
                    "productID": "cWGtlAnhcsVViidg",
                    "quantity": 2
                  }
                  ],
                  "totalAmount": 790.01,
                  "seller":"ofqRrrA7womznAeY",
                  "updatedAt":"2024-01-07T17:15:24.625Z",
                  "orderDate": "2023-12-01T14:30:00Z",
                  "status": "placed"
                }]
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/order/{orderID}": {
      "get": {
        "tags": [
          "order"
        ],
        "summary": "Get Order by OrderID",
        "operationId": "getOrderByID",
        "responses": {
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Order returned",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "buyer": "cWGtlAnhcsVViidg",
                  "articles": [{
                    "productID": "cWGtlAnhcsVViidg",
                    "quantity": 2
                  }
                  ],
                  "totalAmount": 790.01,
                  "seller":"ofqRrrA7womznAeY",
                  "updatedAt":"2024-01-07T17:15:24.625Z",
                  "orderDate": "2023-12-01T14:30:00Z",
                  "status": "placed"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "order"
        ],
        "summary": "Patch Order by OrderID",
        "operationId": "patchOrder",
        "requestBody": {
          "$ref": "#/components/requestBodies/OrderPatch"
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Order returned",
            "content": {
              "application/json": {
                "example": {
                  "_id": "cWGtlAnhcsVViidg",
                  "buyer": "cWGtlAnhcsVViidg",
                  "articles": [{
                    "productID": "cWGtlAnhcsVViidg",
                    "quantity": 2
                  }
                  ],
                  "totalAmount": 790.01,
                  "seller":"ofqRrrA7womznAeY",
                  "updatedAt":"2024-01-07T17:15:24.625Z",
                  "orderDate": "2023-12-01T14:30:00Z",
                  "status": "placed"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/cart": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Create a new Shopping-Cart",
        "operationId": "createCart",
        "requestBody": {
          "$ref": "#/components/requestBodies/Cart"
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "201": {
            "description": "Cart created",
            "content": {
              "application/json": {
                "example": {
                  "articles": [
                    {
                      "productId": "jwxq65KLiuM1C2z",
                      "quantity": 2
                    },
                    {
                      "productId": "jwxq65KLiuM1C2zz",
                      "quantity": 2
                    }
                  ],
                  "buyer": "pyh0X1ZFKfI04HMX",
                  "seller": "pyh0X1ZFKfI04HMX",
                  "totalAmount": 4000.04,
                  "createdAt": "2024-01-07T17:15:15.133Z",
                  "updatedAt": "2024-01-07T17:15:16.469Z",
                  "_id": "ObNRuplzhb0HPUhp"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/cart/{cartID}": {
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Get Shopping-Cart by cartID",
        "operationId": "getCartByID",
        "requestBody": {
          "$ref": "#/components/requestBodies/Cart"
        },
        "responses": {
          "404": {
            "description": "Cart not found"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Cart returned",
            "content": {
              "application/json": {
                "example": {
                  "articles": [
                    {
                      "productId": "jwxq65KLiuM1C2z",
                      "quantity": 2
                    },
                    {
                      "productId": "jwxq65KLiuM1C2zz",
                      "quantity": 2
                    }
                  ],
                  "buyer": "pyh0X1ZFKfI04HMX",
                  "seller": "pyh0X1ZFKfI04HMX",
                  "totalAmount": 4000.04,
                  "createdAt": "2024-01-07T17:15:15.133Z",
                  "updatedAt": "2024-01-07T17:15:16.469Z",
                  "_id": "ObNRuplzhb0HPUhp"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "cart"
        ],
        "summary": "Update Shopping-Cart",
        "operationId": "updateCartByID",
        "requestBody": {
          "$ref": "#/components/requestBodies/Cart"
        },
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Cart updated",
            "content": {
              "application/json": {
                "example": {
                  "articles": [
                    {
                      "productId": "jwxq65KLiuM1C2z",
                      "quantity": 2
                    },
                    {
                      "productId": "jwxq65KLiuM1C2zz",
                      "quantity": 2
                    }
                  ],
                  "buyer": "pyh0X1ZFKfI04HMX",
                  "seller": "pyh0X1ZFKfI04HMX",
                  "totalAmount": 4000.04,
                  "createdAt": "2024-01-07T17:15:15.133Z",
                  "updatedAt": "2024-01-07T17:15:16.469Z",
                  "_id": "ObNRuplzhb0HPUhp"
                }
              }
            }
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "cart"
        ],
        "summary": "Delete Shopping-Cart",
        "operationId": "deleteCart",
        "requestBody": {
          "$ref": "#/components/requestBodies/Buyer"
        },
        "responses": {
          "401": {
            "description": "Authorization Header missing.\n Wrong Password"
          },
          "403": {
            "description": "Wrong Username"
          },
          "500": {
            "description": "Internal Server Error."
          },
          "200": {
            "description": "Cart deleted"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "Order": {
        "type": "object",
        "required": [
          "buyer",
          "articles",
          "status"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "buyer": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "seller": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "articles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productID": {
                  "type": "string",
                  "format": "BSON",
                  "example": "cWGtlAnhcsVViidg"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "example": 3
                }
              }
            },
            "example": [
              {"productID": "cWGtlAnhcsVViidg", "quantity": 2},
              {"productID": "xyz123", "quantity": 1}
            ]
          },
          "totalAmount": {
            "type": "number",
            "format": "double",
            "multipleOf": 0.01,
            "readOnly": true
          },
          "orderDate": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "status": {
            "type": "string",
            "description": "Order Status",
            "enum": [
              "placed","payed", "shipped", "delivered", "canceled"
            ]
          }
        }
      },
      "OrderPatch": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "Order Status",
            "enum": [
              "placed","payed", "shipped", "delivered", "canceled"
            ]
          }
        }
      },
...
      "Buyer": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "iban",
          "address",
          "zipCode",
          "city"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "BSON",
            "readOnly": true
          },
          "password": {
            "type": "string",
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=!]).{8,}$",
            "example": "Test123#"
          },
          "firstName": {
            "type": "string",
            "example": "Tarek"
          },
          "lastName": {
            "type": "string",
            "example": "Bürner"
          },
          "email": {
            "type": "string",
            "example": "test@test.de"
          },
          "address": {
            "type": "string",
            "example": "Leiser Weg 12"
          },
          "zipCode": {
            "type": "number",
            "pattern": "5 digits, existing zipCode",
            "example": 70569
          },
          "city": {
            "type": "string",
            "pattern": "must match the zipCode",
            "example": "Stuttgart"
          }
        }
      },
...
      "Cart": {
        "type": "object",
        "required": "articles",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productID": {
                  "type": "string",
                  "format": "BSON",
                  "example": "cWGtlAnhcsVViidg"
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 1,
                  "example": 3
                }
              }
            },
            "example": [
              {"productID": "cWGtlAnhcsVViidg", "quantity": 2},
              {"productID": "xyz123", "quantity": 1}
            ]
          }
        }
      },
...
    },
    "requestBodies": {
      "Buyer": {
        "description": "Buyer, that should be registered",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Buyer"
            }
          }
        },
        "required": true
      },
...
      "Order": {
        "description": "Seller, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "OrderPatch": {
        "description": "Seller, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OrderPatch"
            }
          }
        }
      },
      "Cart": {
        "description": "Seller, that should be updated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Cart"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
...
  }
}